feat: add Competition and Attendance report pages with export functionality
