+++ src/App.jsx
@@ -29,6 +29,8 @@ const Attendance = lazy(() => import("./pages/Student/Attendance"));
 const Thread = lazy(() => import("./pages/Thread/Thread"));
 const ThreadWindow = lazy(() => import("./pages/Thread/ThreadWindow"));
 const Report = lazy(() => import("./pages/Report/Report"));
+const CompetitionReportPage = lazy(() => import("./pages/Report/CompetitionReportPage"));
+const AttendanceReportPage = lazy(() => import("./pages/Report/AttendanceReportPage"));
 const ThreadReports = lazy(() => import("./pages/ThreadReports/ThreadReports"));
 const ThreadReportsByMentor = lazy(() => import("./pages/ThreadReports/ThreadReportsByMentor"));
 const ThreadReportsByStudent = lazy(() => import("./pages/ThreadReports/ThreadReportsByStudent"));
@@ -493,6 +495,22 @@ function App() {
                     element={
                       <ProtectedRouteWrapper allowedRoles={["admin", "hod", "director", "strcoordinator"]}>
                         <LazyLoadWrapper component={Report} />
+                      </ProtectedRouteWrapper>
+                    }
+                  />
+                  <Route
+                    path="/report/competition"
+                    element={
+                      <ProtectedRouteWrapper allowedRoles={["admin", "hod", "director", "strcoordinator"]}>
+                        <LazyLoadWrapper component={CompetitionReportPage} />
+                      </ProtectedRouteWrapper>
+                    }
+                  />
+                  <Route
+                    path="/report/attendance"
+                    element={
+                      <ProtectedRouteWrapper allowedRoles={["admin", "hod", "director", "strcoordinator"]}>
+                        <LazyLoadWrapper component={AttendanceReportPage} />
                       </ProtectedRouteWrapper>
                     }
                   />
